pyproject.toml
[tool.poetry]
name = "chatbot-tlcn"
version = "0.1.0"
description = ""
authors = ["Nguyễn Tấn Phát <phatnguyen.041203@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
nbconvert = "^7.16.4"
pydantic = "^2.9.2"
psycopg2 = "^2.9.9"
jsonlines = "^4.0.0"
html5lib = "^1.1"
beautifulsoup4 = "^4.12.3"
google-generativeai = "^0.8.3"
pandas = "^2.2.3"
aiohttp = "^3.10.10"
openpyxl = "^3.1.5"
pgvector = "^0.3.5"
chainlit = "^1.3.1"
openai = "1.53"
google-api-python-client = "^2.154.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
phonenumbers = "^8.13.50"
email-validator = "^2.2.0"
redis = "^5.2.0"
rq = "^2.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
notebook = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
